requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
packaging==21.3
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
spotifyapi==0.1.1
spotipy==2.21.0
sqlparse==0.4.3
urllib3==1.26.12
wrapt==1.14.1
